Initial commit.
And so it begins, another project. This one generates (possibly) secure
passwords. Well, right now, it just generates a whole lotta text and no
passwords.

This lays out a basic code structure and accepts some arguments. The
core output format has been set up here. Now we can begin filling in the
actual password generation.

I decided upon the `pywrdgen` name because I don't see anything else
using it, and it felt unique. Not totally in love with it, but maybe
it'll grow on me.