Begin `Password` class. Import it.
Started work on the `Password` class here. This class will contain all
of the attributes used during generation of the password, as well as the
methods used to access or generate it. There are stub methods to give us
a guideline of what to develop, but no solid implementation has been
started yet.