Clean up the code a bit, make life easier
Taking a good look over the code, and a `flake8` run or three, we're a
lot more cleaned up than before. Much of this commit consists of
modifications to whitespace such as newlines, or indentation. We also
added a few comments to help us navigate the code slightly easier. The
only functional change to the `pywrdgen`'s operation is the updating of
the title displayed in the info blurb.