Flesh out the `Password` class a little more.
The star of the show here is probably going to be the skeleton code laid
out in `Password.generate`. Really just kinda provides a base guideline
on where we might start filling in the proper generate function. The
current 'password' it assembles is a string specifying that instance's
attributes.

Which also brings up the inclusion of our `Password` attributes into the
class. These are grabbed during `__init__` from the flags passed into
`__main__.gen` through the context. These attributes are used to
describe the rules followed when generating it.

Finally, make note of an issue I noticed. When we don't explicitly
generate the password in `__main__.gen`, the access of it later fails in
the loop to output the generated passwords.