Generate some passwords
In the most basic sense, we've set out what we came to accomplish.
Random password generation. However, this is not finished nor half as
secure as it should be.

The password is currently generated by assembling a series of character
sets, based on the enabled CLI flags. That list of character sets is
then passed through `random.sample()` tomix it up and grab `--length`
number of characters. There is no guarantee that any number of
characters from each character set is included. This can lower password
strength by decreasing the variety of characters that may be included
for any given run.