`Password` getter properly working now
Turns out the getter was not working right after all. The call to
`Password.generate` was not properly going through. Still not entirely
sure what it *was* doing, not the right thing I know that much.

It seems to work now though. No explicit need to call
`Password.generate` on new objects of the type.