Partial hexadecimal password output mode
Partial implementation of a hexadecimal output mode. In that mode, a
generated password will be comprised only of the hexadecimal digits
[A-Z0-9].

This is kind of working. It highlights an issue with our
`Password.generate` method. Because was call `random.sample` on the
character set, it is actually pull values *out* of that list. No single
character may be pulled more than once in this case. And if the password
length is longer than the number of items in the character set,
exceptions get thrown. I'm gonna have to work on this.