Refactor `Password.generate` a touch
Rather than grabbing a `random.sample` of individual characters in our
character set, grab them on at a time. Looping over the set, grabbing
one at random for every character in the length of the password being
generated.

This fixes the issue where passwords could not be longer than the
character set they are made up of.