Enable `--upper` flag in hex password mode.
Small tweak to `Password.build_character_set so that the presence of the
`--upper` flag will enable the hexadecimal password generated to use
only uppercase alphabetic characters. The absence of that flag results
in lowercase alphabetic character output. The `--upper` flag overrides
the presences of an `--alpha` flag in the same command, regardless of
position.