Fix errors in `setup.py`
Doing a final test of the package with an upload to the test instance of
PyPi, I noticed a couple problems. For one, we passed an incorrect
argument for the long description type. Causing the Markdown to be
rendered as reStructuredText which was not was working.

Also remove the `py_modules` parameter. It was mistakenly copied over
from another project's `setup.py` thinking without knowing what it did.