Stub `Password.strength` property
Add the stub of a strength property into the `Password` class. This will
return an estimate of the password strength, generating a password if
one has not already been.

This kinda blends with the previous commit. Where I accidentally `git
commit -a` instead of `git commit`. Small changes, so no big deal. Still
would have liked the separation though.