Utilize `get` instead of direct access
This is a safety check we probably should have been using earlier, but
didn't need to. Now it's a bit more important. Because we use one
`Password` class that can operate in two different modes, called from
two different commands in `__main__`.. It is now important to check for
the existence of a key we want to grab into our attributes from
`**kwargs`.

This is because if we enter `Password.__init__` from another `click`
command, were all of the arguments are not specified on the command
line, we will hit a `KeyError`. At least, if we were to directly access
the argument with a `kwargs['argument']` type statement. If a given key
does not exist in the `kwargs` dictionary, it will be assigned a default
of `None`.