Stub `Password.generate_passphrase` method
Down the road, this will handle generating a passphrase. I chose to
include this into a separate function because while similar to
`generate`, the behavior feels different enough to warrant it.

Basic flow of the function has been laid out in the method comments.
Some initial experimenting in the REPL has given me a good feel on how I
should try to implement this.