Rough draft passphrase generator
Add the first iteration of our lovely new passphrase generator. This
commit does several things. I'll try to briefly break them down, without
as much of an essay as usual. In approximate order of execution, we..

  - Pass the `--debug` flag value into `Password`, this allows us to
    force passphrase generation without having the command made yet.
  - Add attributes to `Password` for determining the EFF Wordlist path.

  - `Password.generate_passphrase`
    - Load the word list (on demand loading)
    - For each of the total words requested in the passphrase, loop
      - Roll five dice, six sides each. `_roll_dice` is a generator, so
	we iterate over the returned roll results.
        - Concatenate string dice results into a 5-character index.
      - Iterate over `_word_list` and return the line which matches our
	previously generated word index. This is the list comprehension.
	Should only return one result.
      - Split the line on '\t' to the left, and a '\n' to the right
        - This can be more readible. Feels hacky right now.
      - We now have the word from the EFF Wordlist that corresponds to
	the result of our five dice rolls!
      - Append that word onto the password. Deduct one from the
	`word_remaining` counter. Continue looping.

That all said, this works based on my testing so far. It is a *very*
rough draft. Just enough to get some words off our list. The proper CLI
command, argument specification, and improved passphrases will be coming
shortly. For now, to force passphrase generation, and test this new
feature, pass the `--debug` flag like so:

```
$ pywrdgen --debug gen --count 6 --length 3
$ pywrdgen -d gen -c 6 -l 3
```